+++ src/SFML/System/Win32/Platform.hpp
@@ -47,18 +47,18 @@ public :
     ////////////////////////////////////////////////////////////
     /// \brief Get the current system time
     ///
-    /// \return System time, in seconds
+    /// \return System time, in milliseconds
     ///
     ////////////////////////////////////////////////////////////
-    static double GetSystemTime();
+    static Uint64 GetSystemTime();
 
     ////////////////////////////////////////////////////////////
-    /// \brief Suspend the execution of the current thread for a specified time
+    /// \brief Suspend the execution of the current thread for a specified duration
     ///
-    /// \param time Time to sleep, in seconds
+    /// \param time Time to sleep, in milliseconds
     ///
     ////////////////////////////////////////////////////////////
-    static void Sleep(float time);
+    static void Sleep(Uint32 time);
 };
 
 } // namespace priv
